refactor: divide code into modules
